+++ Sources/code/application/Models/Cart/PromotionOfferCartAppModel.swift
@@ -28,6 +28,8 @@ public extension ApplicationClient.Cart {
         
         public var description: String?
         
+        public var isBankOffer: Bool?
+        
 
         public enum CodingKeys: String, CodingKey {
             
@@ -51,9 +53,11 @@ public extension ApplicationClient.Cart {
             
             case description = "description"
             
+            case isBankOffer = "is_bank_offer"
+            
         }
 
-        public init(buyRules: [String: Any]? = nil, description: String? = nil, discountRules: [[String: Any]]? = nil, freeGiftItems: [FreeGiftItems]? = nil, id: String? = nil, offerText: String? = nil, promotionGroup: String? = nil, promotionName: String? = nil, promotionType: String? = nil, validTill: String? = nil) {
+        public init(buyRules: [String: Any]? = nil, description: String? = nil, discountRules: [[String: Any]]? = nil, freeGiftItems: [FreeGiftItems]? = nil, id: String? = nil, isBankOffer: Bool? = nil, offerText: String? = nil, promotionGroup: String? = nil, promotionName: String? = nil, promotionType: String? = nil, validTill: String? = nil) {
             
             self.id = id
             
@@ -74,6 +78,8 @@ public extension ApplicationClient.Cart {
             self.freeGiftItems = freeGiftItems
             
             self.description = description
+            
+            self.isBankOffer = isBankOffer
             
         }
 
@@ -200,6 +206,18 @@ public extension ApplicationClient.Cart {
             }
             
             
+            
+            do {
+                isBankOffer = try container.decode(Bool.self, forKey: .isBankOffer)
+            
+            } catch DecodingError.typeMismatch(let type, let context) {
+                print("Type '\(type)' mismatch:", context.debugDescription)
+                print("codingPath:", context.codingPath)
+            } catch {
+                
+            }
+            
+            
         }
         
         public func encode(to encoder: Encoder) throws {
@@ -243,6 +261,10 @@ public extension ApplicationClient.Cart {
             
             
             try? container.encodeIfPresent(description, forKey: .description)
+            
+            
+            
+            try? container.encodeIfPresent(isBankOffer, forKey: .isBankOffer)
             
             
         }